plastic_rice_detection/requirements.txt
tensorflow>=2.8.0
streamlit
opencv-python
numpy
Pillow
scikit-learn 